Remove libsodium dependency
I had moved things to libsodium with the thought of eventually
statically linking to it, having forgotten that libsodium is a fairly
large library to statically link to. Now knowing that we're going to be
dynamically linking to a standard crypto lib for randbytes and constant
time memcmp, OpenSSL is the clear choice due to its strong chance of
already being installed on machines using this library.