Update warn_unused_result attributes
For now I'm only going to use this attribute for _gensalt() where it's
easy to overlook the importance of checking the return value (as many
crypto libraries pre-initialize their RNG in order for randombytes to
not have a chance of erroring). Going to do my best to keep header files
clean.